Moved class defs into shared library
